[Uploader] Move file preview into own component
