+++ app/javascript/src/javascripts/tag_editor.vue
@@ -29,7 +29,7 @@
 <script>
   import Vue from 'vue';
   import relatedTags from './uploader_related.vue';
-  import tagPreview from './uploader_preview.vue';
+  import tagPreview from './uploader_tag_preview.vue';
   import Post from './posts';
   import Autocomplete from "./autocomplete.js.erb";
   import Utility from "./utility.js";
