+++ app/javascript/src/javascripts/uploader_tag_preview.vue
